# openapi.yaml
openapi: 3.1.0
info:
  title: WordPress Publishing Bridge
  version: 1.2.0
servers:
  - url: https://your-bridge.example.com
paths:
  /publish:
    post:
      operationId: publishPage
      summary: Publish HTML content to WordPress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, html]
              properties:
                title:
                  type: string
                html:
                  type: string
                status:
                  type: string
                  enum: [draft, publish, private]
                  default: draft
                content_type:
                  type: string
                  enum: [page, post]
                  default: page
                slug:
                  type: string
                excerpt:
                  type: string
      responses:
        "200":
          description: Publish result

  /publish_transcript:
    post:
      operationId: publishTranscript
      summary: Publish a conversation transcript to WordPress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, messages]
              properties:
                title:
                  type: string
                status:
                  type: string
                  enum: [draft, publish, private]
                  default: draft
                content_type:
                  type: string
                  enum: [page, post]
                  default: page
                slug:
                  type: string
                intro:
                  type: string
                messages:
                  type: array
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      content:
                        type: string
      responses:
        "200":
          description: Publish transcript result